Rename Primitive to PrimitiveWithId and move it to typechecking.rs
